Manejo excepción para eliminación de Department con id inexistente.

+++ src/main/java/com/nico/springbootprueba/service/DepartmentServiceImpl.java
@@ -31,8 +31,10 @@ public class DepartmentServiceImpl implements DepartmentService {
     }
 
     @Override
-    public void deleteDepartmentById(int id) {
-        departmentRepository.deleteById(id);
+    public void deleteDepartmentById(int id) throws DepartmentNotFoundException {
+        if (departmentRepository.existsById(id))
+            departmentRepository.deleteById(id);
+        throw new DepartmentNotFoundException("Department not available");
     }
 
     @Override
